Add initial NTT based polynomial multiplication
This is some very rough work on multiplying two polynomials within Z_17
[x] / (x^4 + 1). While the basic steps are there, the inverse NTT
transform is definitely not finished. I *think* this only works for a 2
layer polynomial ring but at this hour I'm too dumb to figure this out.

The next steps are to generalize the inverse NTT transform and apply
this technique to a multiplication of size 8 polynomials.