Add common file containing reusable components
This file is used to contain some helper functions (pad, reduce_q) and
the classes Goods, NTT. This allows for writing cleaner code samples and
makes testing individual components easier. Every function and class has
been documented and annotated with type hints.

In the future it might make sense to split this into subcomponents
(helpers, Goods, NTT) but for now this is fine.