Add size - 8 NTT based polynomial multiplication
This is a layered size - 8 NTT based polynomial multiplication
implemented in C. The first step towards a complete C implementation.

The functions can be shortened here and there but I think it strikes a
nice balance between readability and simplicity at the moment.

There are a couple of TODO tags that need some attention some time in
the future.