Add 'kem/ntrulpr761' NTT example
This commit adds an iterative and recursive version for performing NTT
based polynomial multiplication of two polynomials for the original NTRU
LPRime 'kem/ntrulpr761' parameter set - in Python. :partying_face:

Choices for variables, parameters and overall approach should be well
defined and explained. For further clarification you might need to refer
to one of the smaller examples which are easier to follow and compute by
hand. The Z_16 [ ] / (x^11 - x - 1) NTT example would be the first
go-to.