Add initial incorrect size - 512 NTT
This commit outlines the structure and declares the values that the 512
NTT has. It computes "poly_one * poly_two % (x^512 - 1) % 6984193" and
compares the result to the known value.

The current implementation is not yet correct, most likely due to the
fact that the used numeric types (uint32_t) are not large enough to
contain the intermediate results. TODO tags have been placed where this
might occur.