Implement Good's and Montgomery multiplication
This commit includes a number of changes, which would take me too much
time to separate nicely. They include:

1. Implemented Good's forward and inverse permutations. This allows us
to deconstruct our 'clunky' NTT into 3 size-512 NTTs after zero-padding
the integer arrays.

2. Implemented Montgomery multiplication. This allows us to do fast
modular multiplication. This change requires us to ensure that one of
the multiplicands is in the Montgomery domain. The roots and a number of
constants have therefore changed as a result of this.

3. Implemented additional thorough documentation to explain all this in
detail. The code should be sufficiently easy to follow what happens,
where and why.