Initial ugly kem/ntrulpr761 implementation
It might be ugly, it might be a bit clunky in places, but it works.

This commit implements the initial kem/ntrulpr761 implementation.
Nothing special has been added, the implementation mostly follows what
was already there in Python, but now in C.

This definitely needs some cleaning up but that's a job for tomorrow-me.
There are also a couple of TODO tags that need my attention. Hi me from
the future! :wave: