Update to a proper kem/ntrulpr761 known value test
This commit updates poly_two to be a small/short polynomial similar to
how it would be used in NTRU LPRime. Instead of taking elements from {0,
1} we now include the correct amount of elements from {-1, 0, 1}.

This revealed a new bug which requires us to ensure that the
intermediate result is correct in the integer domain before the final
reduction mod 4591.

:beetle: . :tennis: