Express numeric types explicitly
This commit adds two explicit type casts in util.c such that the intent
is even clearer with regards to the numeric types. Without this the
compiler would complain about the conversions (from int64_t to int32_t)
as they may change the values unintentionally.

We don't want the compiler to complain.