Include additional flags that need to be called explicitly
This commit adds the '-fomit-frame-pointer' compiler flag to the build
process. This instructs the compiler to not store stack frame pointers
during function calls, if a function does not need it.

This coincidentally allows us to use R29 as a general purpose register,
and makes reading generated assembler code a lot easier.