Create vectorized _asimd_mul_red macro
This commit adds a vectorized Montgomery reduction macro called
`_asimd_mul_red`. The docstring right above contains some additional
information as to why it works and why it has been implemented as-is.

This documentation will most likely get moved somewhere else down the
line as this is more suitable to include in the paper.