Create __asm_ntt_forward_layer macro
This macro implements a forward butterfly and can be reused by any other
layer that has length >= 4. This limitation is due to the __asimd_
macro's and the fact that they take 4 32-bit integer coefficients
simultaneously.

If this condition is not met, execution of the macro will produce a
segmentation fault.