Implement initial forward_layer_8 forward_layer_9
This commit implements an initial *very* ugly implementation of
__asm_ntt_forward_layer_8 and __asm_ntt_forward_layer_9. This is merely
to complete each forward layer in assembly.

This definitely needs some additional work as the current implementation
for layers 8 and 9 is slower than what the compiler auto generates. This
is due to the fact that the SIMD registers and instructions are used for
2, respectively 1 integer coefficient arithmetic instead of utilizing
the full 128 bits.