Sketch outline of vectorized approach - 2
This commit reformats the C sources for the inverse NTT layers 8 and 9
to sketch what we are trying to accomplish. This makes it easier to see
how to build the ASIMD code.

From inverse_layer_7 onwards (excluding the final layer) it will be
relatively straightforward as we are once again able to use chunks of 4
32-bit integer coefficients (length >= 4).