Add main512 NTT example
This commit adds main512.c and main512.h which can be used to perform
and benchmark a single size 512 NTT application. It uses the same
functions and sources as main761 but focuses only on a single NTT.

This can be used to calculate Zx(F) * Zx(G) % (x^512 - 1) % 6984193.

Similarly to main761 it includes a SPEED switch which can be used to
either enable or suppress printing debug information.